Reimplement per-operator selective build
This uses the technique proposed by smessmer in D16451848 to selectively
register operators without codegen.  See the Note inside for more
details.

This PR has feature parity with the old selective build apparatus:
it can whitelist schema def()s, impl()s, and on a per dispatch key
basis.  It has expanded dispatch key whitelisting, whereas previously
manually written registrations were not whitelisted at all.  (This
means we may be dropping dispatch keys where we weren't previously!)

With this new implementation, we remove code generated whitelisting,
generate to separate files

- Revert "[pytorch] move force schema registration output into a separate file (#36284)"
  This reverts commit 126d00c8ddcca996a5045f6a6c59cf86c120936c.
- Revert "[pytorch] force c10 schema registration for custom build"
  This reverts commit 61b680c01250c420603ae2937a2f2f16adebc475.
- Revert "[pytorch] aten codegen to filter backends for default mobile build"
  This reverts commit 6792bafa7258e66cf66f738fac6ab8316c4bc0f8
- Revert "[pytorch] codegen flags to whitelist op registrations / generate to separate files (#32451)"
  This reverts commit 69283388caa1fabe695ca0209ea03bed830de9be.

The new plan on record is that we will control whether or not
we should register an operator or not by a non-codegenned mechanism
(constexpr conditional).

This PR doesn't actually implement this new mechanism yet (for ease
of review), which means it will break all of our selective builds.
I intend to squash it with a future PR later.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 5f1f662bd31871b76818617a396ab4e016389c6d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39401